timeout

+++ RandomShuffleExam/CodeWorker.cpp
@@ -1,4 +1,4 @@
-/********************************************************************************
+﻿/********************************************************************************
 *   Copyright (C) 2019 by Bach Nguyen Sy                                       *
 *   Unauthorized copying of this file via any medium is strictly prohibited    *
 *                                                                              *
@@ -15,6 +15,7 @@
 */
 
 #include "CodeWorker.h"
+#include "Vietnamese.h"
 #include <QDir>
 #include <QProcess>
 #include <QVector>
@@ -32,10 +33,14 @@ void CodeWorker::process()
 {
 	const auto input = QString("\"%1\"").arg(m_exam);
 	const auto output = QString("\"%1\"").arg(m_saveDir);
-	auto cmd = "7z.exe x " + input + " -o" + output;
+	auto cmd = "bin\\7z.exe x " + input + " -o" + output;
 	QProcess process(this);
 	process.start(cmd);
-	process.waitForFinished();
+	if(!process.waitForFinished(300000))
+	{
+		emit error(Vietnamese::str(L"E: Quá thời gian"));
+		return;
+	}
 
 	const auto baseName = QFileInfo(m_exam).baseName();
 	QDir dir(m_saveDir);
@@ -43,9 +48,13 @@ void CodeWorker::process()
 
 	const auto zipName = QString(R"("%1\%2.zip")").arg(m_saveDir).arg(m_code);
 	const auto fileName = QString(R"("%1\%2\*")").arg(m_saveDir).arg(m_code);
-	cmd = "7z.exe a -tzip " + zipName + " " + fileName;
+	cmd = "bin\\7z.exe a -tzip " + zipName + " " + fileName;
 	process.start(cmd);
-	process.waitForFinished();
+	if (!process.waitForFinished(300000))
+	{
+		emit error(Vietnamese::str(L"E: Quá thời gian"));
+		return;
+	}
 
 	dir.cd(m_code);
 	dir.removeRecursively();
